Convert BaseSnapExecutor test
